// src/types/subscription.interface.ts
export interface SubscriptionPlanFeatures {
    maxTokens: number;
    maxStaticDataSourceCount: number;
    maxDynamicDataSourceCount: number;
    maxUserAgentCount: number;
    maxImageInterpretationCount: number;
    maxTextToImageCount: number;
    maxTextToVideoCount: number;
    speechToTextCount: {
        maxCount?: number;
        unlimited: boolean;
    }
}

export interface SubscriptionFeature {
    name: string;
    price: number;
    features: SubscriptionPlanFeatures;
    additionalPerks: string;
    subscriptionPlan: string;
    interval: PlanDuration;
    planType: PlanType;
    priority: number;
    livemode: boolean;
}
// ...
export interface SubscriptionPlan {
    _id: string;
    interval: string;
    livemode: boolean;
    name: string;
    planType: string;
    price: number;
    priority: number;
    stripeProductId: string;
    stripePriceId: string;
    subscriptionFeature: SubscriptionFeature;
}